requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
defusedxml==0.7.1
Django==3.2.8
django-allauth==0.45.0
django-ckeditor==6.1.0
django-crispy-forms==1.13.0
django-environ==0.8.1
django-js-asset==1.2.2
django-livereload-server==0.3.2
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.3
oauthlib==3.1.1
Pillow==8.4.0
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
tornado==6.1
urllib3==1.26.7
whitenoise==5.3.0